Section 3 Draw the badge symbol
The Landmarks badge has a custom insignia in its center that’s based on the mountain that appears in the Landmarks app icon.
The mountain symbol consists of two shapes: one that represents a snowcap at the peak, and the other that represents vegetation along the approach. You’ll draw them using two partially triangular shapes that are set apart by a small gap.