App Design and Layout Composing complex interfaces
The category view for Landmarks shows a vertically scrolling list of horizontally scrolling landmarks. As you build this view and connect it to your existing views, you’ll explore how composed views can adapt to different device sizes and orientations.
Follow the steps to build this project, or download the finished project to explore on your own.